Use separate components for single filter menus
